Moved script to its own repository as it no longer represents the code it was based upon
The fork as it originally existed as michealespinola/plexupdate is nothing like the original and although a fork in spirit, it no longer represents the code it was based upon. Although some aspects of the logic remain, the code is almost entirely unique and bears no semblance to martinorob/plexupdate. Its code is unique, as is its support requirements - so it requires its own repository with its own issues tracker.