Fix check whether update was successful
In commit 2b4d78f the current and new versions were trimmed of their build number, but the version checked after installation to see if it succeeded was missed, breaking the update check. If it failed the script would still think the update succeeded because 1.22.2.4282-6000 (non-trimmed version) is "greater than" 1.22.2.4282 (trimmed version).