Codecs name file for codec download/update feature
This file contains the core codec names used in the file naming conventions for decoder and encoder shared library files (.so) used by Plex on the Synology platform. These names are processed and merged with other localized Plex Media Server data as well as Synology specs to build a full URL to download the codec file with.